ai_cmd/requirements.txt
# AI量化交易助手依赖
ssquant>=0.2.3
openai>=1.0.0
tiktoken>=0.5.0
rich>=10.0.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.1
python-dotenv>=0.19.0
